# src/api/openapi.yaml
openapi: 3.1.0
info:
  title: Cashlint API
  version: 1.0.0
  description: Accrual-based Accounting App
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    Vendor:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        balance:
          type: number
          format: decimal
          description: Current subsidiary balance (Accounts Payable)
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - balance
        - createdAt
    VendorBill:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
        vendorId:
          type: string
          format: uuid
        billNumber:
          type: string
        amount:
          type: number
          format: decimal
        date:
          type: string
          format: date-time
        dueDate:
          type: string
          format: date-time
        description:
          type: string
        status:
          type: string
          enum: [Draft, Recorded, PartiallyPaid, Paid]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - vendorId
        - amount
        - date
        - status
        - createdAt

    # Purchasing Context (additional schemas)
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            type:
              type: string
              enum: [DomainFailure, InfrastructureFailure, ApplicationFailure]
            subtype:
              type: string
            message:
              type: string
          required:
            - type
            - subtype
            - message
...
security:
  - cookieAuth: []
paths:
...
  /api/purchasing/vendors:
    post:
      summary: Create a new vendor
      description: Create a vendor record for the authenticated user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  format: uuid
                name:
                  type: string
                  minLength: 1
                  maxLength: 200
                email:
                  type: string
                  format: email
              required:
                - userId
                - name
      responses:
        '201':
          description: Vendor created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  vendor:
                    $ref: '#/components/schemas/Vendor'
                  message:
                    type: string
        '400':
          $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: List vendors
      description: Retrieve vendors for the authenticated user.
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: skip
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
        - name: take
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: List of vendors
          content:
            application/json:
              schema:
                type: object
                properties:
                  vendors:
                    type: array
                    items:
                      $ref: '#/components/schemas/Vendor'
                  count:
                    type: integer
        '400':
          $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/schemas/ErrorResponse'
  /api/purchasing/vendor-bills:
    post:
      summary: Record a vendor bill
      description: Record a bill from a vendor for goods/services received.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  format: uuid
                vendorId:
                  type: string
                  format: uuid
                billNumber:
                  type: string
                amount:
                  type: number
                  minimum: 0.01
                date:
                  type: string
                  format: date-time
                dueDate:
                  type: string
                  format: date-time
                description:
                  type: string
              required:
                - userId
                - vendorId
                - amount
                - date
      responses:
        '201':
          description: Vendor bill recorded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  bill:
                    $ref: '#/components/schemas/VendorBill'
                  message:
                    type: string
        '400':
          $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Vendor not found
        '500':
          $ref: '#/components/schemas/ErrorResponse'
  /api/purchasing/loan-payments:
    post:
      summary: Record a loan payment
      description: Record a loan payment with principal and interest split.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  format: uuid
                vendorId:
                  type: string
                  format: uuid
                principalAmount:
                  type: number
                  minimum: 0
                interestAmount:
                  type: number
                  minimum: 0
                date:
                  type: string
                  format: date-time
                description:
                  type: string
              required:
                - userId
                - vendorId
                - principalAmount
                - interestAmount
                - date
      responses:
        '201':
          description: Loan payment recorded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  loanPayment:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                  message:
                    type: string
        '400':
          $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Vendor not found
        '500':
          $ref: '#/components/schemas/ErrorResponse'
  /api/purchasing/cash-expenses:
    post:
      summary: Record a cash expense
      description: Record a one-time cash expense (money out).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  format: uuid
                vendorId:
                  type: string
                  format: uuid
                amount:
                  type: number
                  minimum: 0.01
                date:
                  type: string
                  format: date-time
                expenseCategory:
                  type: string
                description:
                  type: string
              required:
                - userId
                - amount
                - date
                - expenseCategory
      responses:
        '201':
          description: Cash expense recorded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  cashExpense:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                  message:
                    type: string
        '400':
          $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/schemas/ErrorResponse'

  # PeriodClose Context
...